Basic input handling. PlayerCtrlSys handles camera and movement
Former-commit-id: 14fb9288c463d7eb9d12d6183f926e7ad81593d0

+++ include/events/ViewChangedEvent.hpp
@@ -8,6 +8,6 @@ namespace sw {
     struct ViewChangedEvent {
         ViewChangedEvent(int x, int y) : delta_up_(y), delta_right_(x) {}
 
-        int delta_up_, delta_right_;
+        float delta_up_, delta_right_;
     };
 }
